update payment vietqr

+++ src/main/java/com/example/demo/Entity/OrderItem.java
@@ -21,8 +21,9 @@ public class OrderItem {
     private Long bookId;
 
     @Column(name = "quantity")
-    private Long quantity;
+    private Integer quantity;
 
     @Column(name = "price")
     private BigDecimal price;
+
 }
